finance_advisor/frontend/requirements.txt
streamlit
google-cloud-aiplatform[adk,agent_engines]
python-dotenv
google-generativeai
litellm
pydantic
requests
streamlit-local-storage
